requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyQt6>=6.5.0
fast-json-repair>=0.1.0
libcst
tiktoken
docker
modelcontextprotocol
mcp
PyYAML
anthropic
pytest
sseclient
beautifulsoup4
lxml
sentence-transformers 
chromadb 
langchain 
langchain-community 
